+++ documentation/manual/working/javaGuide/main/upload/JavaFileUpload.md
@@ -3,25 +3,23 @@
 
 ## Uploading files in a form using `multipart/form-data`
 
-The standard way to upload files in a web application is to use a form with a special `multipart/form-data` encoding, which allows mixing of standard form data with file attachments. Please note: the HTTP method for the form has to be POST (not GET).
+The standard way to upload files in a web application is to use a form with a special `multipart/form-data` encoding, which lets you mix standard form data with file attachment data.
+
+> **Note:** The HTTP method used to submit the form must be `POST` (not `GET`).
 
 Start by writing an HTML form:
 
-```
-@form(action = routes.Application.upload, 'enctype -> "multipart/form-data") {
+@[file-upload-form](code/javaguide/fileupload/views/uploadForm.scala.html)
 
-    <input type="file" name="picture">
+Now define the `upload` action:
 
-    <p>
-        <input type="submit">
-    </p>
+@[syncUpload](code/javaguide/fileupload/controllers/HomeController.java)
 
-}
-```
+The [`getRef()`](api/java/play/mvc/Http.MultipartFormData.FilePart.html#getRef--) method gives you a reference to a [`TemporaryFile`](api/java/play/libs/Files.TemporaryFile.html). This is the default way Play handles file uploads.
 
-Now let’s define the `upload` action:
+At last, add a `POST` router
 
-@[syncUpload](code/JavaFileUpload.java)
+@[application-upload-routes](code/javaguide.upload.fileupload.routes)
 
 ### Testing the file upload
 
@@ -50,7 +48,7 @@ Using a custom file part handler also means that behavior can be injected, so a 
 
 ## Cleaning up temporary files
 
-Uploading files uses a [`TemporaryFile`](api/java/play/libs/Files.TemporaryFile.html) API which relies on storing files in a temporary filesystem.  All [`TemporaryFile`](api/java/play/libs/Files.TemporaryFile.html) references come from a [`TemporaryFileCreator`](api/java/play/libs/Files.TemporaryFileCreator.html) trait, and the implementation can be swapped out as necessary, and there's now an [`atomicMoveWithFallback`](api/java/play/libs/Files.TemporaryFile.html#temporaryFileCreator--) method that uses `StandardCopyOption.ATOMIC_MOVE` if available.
+Uploading files uses a [`TemporaryFile`](api/java/play/libs/Files.TemporaryFile.html) API which relies on storing files in a temporary filesystem, accessible through the [`getRef()`](api/java/play/mvc/Http.MultipartFormData.FilePart.html#getRef--) method.  All [`TemporaryFile`](api/java/play/libs/Files.TemporaryFile.html) references come from a [`TemporaryFileCreator`](api/java/play/libs/Files.TemporaryFileCreator.html) trait, and the implementation can be swapped out as necessary, and there's now an [`atomicMoveWithFallback`](api/java/play/libs/Files.TemporaryFile.html#temporaryFileCreator--) method that uses `StandardCopyOption.ATOMIC_MOVE` if available.
 
 Uploading files is an inherently dangerous operation, because unbounded file upload can cause the filesystem to fill up -- as such, the idea behind [`TemporaryFile`](api/java/play/libs/Files.TemporaryFile.html) is that it's only in scope at completion and should be moved out of the temporary file system as soon as possible.  Any temporary files that are not moved are deleted.
