+++ documentation/manual/working/javaGuide/main/upload/code/javaguide.upload.fileupload.routes
@@ -1,0 +1,3 @@
+# #application-upload-routes
+POST  /          controllers.HomeController.upload(request: Request)
+# #application-upload-routes
