+++ documentation/manual/working/javaGuide/main/upload/code/javaguide/fileupload/controllers/HomeController.java
@@ -1,0 +1,24 @@
+package javaguide.upload.fileupload.controllers;
+
+//#syncUpload
+import play.libs.Files.TemporaryFile;
+import play.mvc.Controller;
+import play.mvc.Http;
+import play.mvc.Result;
+
+public class HomeController extends Controller {
+
+    public Result upload(Http.Request request) {
+        Http.MultipartFormData<TemporaryFile> body = request.body().asMultipartFormData();
+        Http.MultipartFormData.FilePart<TemporaryFile> picture = body.getFile("picture");
+        if (picture != null) {
+            String fileName = picture.getFilename();
+            String contentType = picture.getContentType();
+            TemporaryFile file = picture.getRef();
+            return ok("File uploaded");
+        } else {
+            return badRequest().flashing("error", "Missing file");
+        }
+    }
+}
+//#syncUpload
