+++ documentation/manual/working/javaGuide/main/upload/code/javaguide/fileupload/views/uploadForm.scala.html
@@ -1,12 +1,12 @@
-@import scalaguide.upload.fileupload.controllers._
+@import javaguide.upload.fileupload.controllers._
 @* #file-upload-form *@
 @helper.form(action = routes.HomeController.upload, 'enctype -> "multipart/form-data") {
-    
+
     <input type="file" name="picture">
-    
+
     <p>
         <input type="submit">
     </p>
-    
+
 }
 @* #file-upload-form *@
